requirements.txt
albumentations>=1.4.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pandas>=2.0.0
pyyaml>=6.0.1
streamlit>=1.33.0
timm>=0.9.0
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.66.0
ultralytics>=8.2.0
